fix(analysis): declare mmap offset as i64 so wasm32 links
The local `extern "C" mmap` in paged_allocate.rs declared `offset: isize`,
which is i32 on wasm32 — a signature mismatch with luaur-common's wasm_libc
`mmap(..., off: i64)` shim that breaks the wasm link of any target pulling in
the typecheck/analysis path. `off_t` is 64-bit on the LP64 native targets too,
so i64 is correct everywhere. Partial fix for #20 (the mmap mismatch; the
undefined strftime/snprintf on the same path remain).

+++ crates/luaur-analysis/src/functions/paged_allocate.rs
@@ -9,7 +9,10 @@ extern "C" {
         prot: c_int,
         flags: c_int,
         fd: c_int,
-        offset: isize,
+        // `off_t` is 64-bit on the LP64 native targets we support, and the wasm
+        // `wasm_libc` shim declares it `i64` too — `isize` would be `i32` on
+        // wasm32, a signature mismatch that breaks the wasm link. Use `i64`.
+        offset: i64,
     ) -> *mut c_void;
 }
